server/requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
greenlet==2.0.2
html5lib==0.999999999
httpcache==0.1.3
idna==3.4
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.2.0
MarkupSafe==2.0.1
numpy==1.24.1
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
PyYAML==6.0
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
singleton-factory==1.0
singleton3==1.0
six==1.3.0
sklearn==0.0.post1
soupsieve==2.3.2.post1
SQLAlchemy==2.0.0
threadpoolctl==3.1.0
typing_extensions==4.4.0
urllib3==1.26.14
webencodings==0.5.1
Werkzeug==2.0.2
